refactor: move store errors into separate file
